Placeholder for RPi <-> Arduino comm interface, RPi side
This is the first platform layer to get staying in place on the e2e systems to maintain integration progress. This communication interface should start being used for everything on the Arduino to be driven by the RPi. We can iterate the comm interface on both boards as its own versioned component.

Currently P0 on Pick and Place Firmware project, please assist Shiivek:

https://github.com/orgs/KoalbyMQP/projects/2/views/1?pane=issue&itemId=87726143